# api/swagger-doc/swagger-one.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "MediFree Backend Api",
      "contact": {}
    },
    "host": "backend-medifree.herokuapp.com",
    "basePath": "/",
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/ciudad/5": {
        "put": {
          "summary": "Borrar una ciudad /ciudad/:idCiudad",
          "tags": [
            "Ciudades"
          ],
          "operationId": "Borrarunaciudad/ciudad/:idCiudad",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/x-www-form-urlencoded"
          ],
          "parameters": [
            {
              "name": "nameCiudad",
              "in": "formData",
              "required": true,
              "type": "string",
              "description": ""
            },
            {
              "name": "idPaisF",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            },
            {
              "name": "isVisible",
              "in": "formData",
              "required": true,
              "type": "integer",
              "format": "int32",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          }
        }
      },
...
    },
...
  }